integrations/showroom-beta/cc24-fullstack/ml-worker/requirements.txt
# CC24 ML Worker - Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Image Processing
Pillow==10.1.0
rembg[gpu]==2.0.50  # Use [cpu] if no GPU available

# Optional: License Plate Detection
# fast-alpr==0.1.0  # Uncomment for plate detection

# Utils
pydantic==2.5.0
